+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## Changelog for 0.10.2
+## Changelog for 0.11.0
+
+### Additions
+
+- **Tooltip Enhancements**
+  - See AH and BZ price directly in the tooltip.
 
 ### Improvements
 
@@ -6,6 +11,7 @@
   - Improved performance
   - Centered the recipe viewer in the middle of the screen
   - Fixed pressing R and U over an item in the inventory not showing the recipe or usage
+  - Improved search calc, you can now use x for multiplication
 - **Chat Enhancements**
   - Pressing the up arrow key no longer selects the chat tab; it now selects the last used command
   - Fixed compact-only consecutive messages not working